+++ packages/vue-components/stories/PrimeVueTailwind.vue
@@ -1,0 +1,72 @@
+;<template>
+  <div class="m-2">
+    <h1 class="text-4xl underline font-medium my-4">Semantic colors</h1>
+
+    <div class="flex">
+      <div class="flex flex-col">
+        <div class="m-2 p-2 rounded-md bg-primary-100">bg-primary-100</div>
+        <div class="m-2 p-2 rounded-md bg-primary-200">bg-primary-200</div>
+        <div class="m-2 p-2 rounded-md bg-primary-300">bg-primary-300</div>
+        <div class="m-2 p-2 rounded-md bg-primary-400">bg-primary-400</div>
+        <div class="m-2 p-2 rounded-md bg-primary-500">bg-primary-500</div>
+        <div class="m-2 p-2 rounded-md bg-primary-600">bg-primary-600</div>
+        <div class="m-2 p-2 rounded-md bg-primary-700">bg-primary-700</div>
+        <div class="m-2 p-2 rounded-md bg-primary-800">bg-primary-800</div>
+        <div class="m-2 p-2 rounded-md bg-primary-900">bg-primary-900</div>
+      </div>
+      <div class="flex flex-col">
+        <div class="m-2 p-2 rounded-md bg-secondary-100">bg-secondary-100</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-200">bg-secondary-200</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-300">bg-secondary-300</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-400">bg-secondary-400</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-500">bg-secondary-500</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-600">bg-secondary-600</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-700">bg-secondary-700</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-800">bg-secondary-800</div>
+        <div class="m-2 p-2 rounded-md bg-secondary-900">bg-secondary-900</div>
+      </div>
+      <div class="flex flex-col">
+        <div class="m-2 p-2 rounded-md bg-success-100">bg-success-100</div>
+        <div class="m-2 p-2 rounded-md bg-success-200">bg-success-200</div>
+        <div class="m-2 p-2 rounded-md bg-success-300">bg-success-300</div>
+        <div class="m-2 p-2 rounded-md bg-success-400">bg-success-400</div>
+        <div class="m-2 p-2 rounded-md bg-success-500">bg-success-500</div>
+        <div class="m-2 p-2 rounded-md bg-success-600">bg-success-600</div>
+        <div class="m-2 p-2 rounded-md bg-success-700">bg-success-700</div>
+        <div class="m-2 p-2 rounded-md bg-success-800">bg-success-800</div>
+        <div class="m-2 p-2 rounded-md bg-success-900">bg-success-900</div>
+      </div>
+      <div class="flex flex-col">
+        <div class="m-2 p-2 rounded-md bg-warning-100">bg-warning-100</div>
+        <div class="m-2 p-2 rounded-md bg-warning-200">bg-warning-200</div>
+        <div class="m-2 p-2 rounded-md bg-warning-300">bg-warning-300</div>
+        <div class="m-2 p-2 rounded-md bg-warning-400">bg-warning-400</div>
+        <div class="m-2 p-2 rounded-md bg-warning-500">bg-warning-500</div>
+        <div class="m-2 p-2 rounded-md bg-warning-600">bg-warning-600</div>
+        <div class="m-2 p-2 rounded-md bg-warning-700">bg-warning-700</div>
+        <div class="m-2 p-2 rounded-md bg-warning-800">bg-warning-800</div>
+        <div class="m-2 p-2 rounded-md bg-warning-900">bg-warning-900</div>
+      </div>
+      <div class="flex flex-col">
+        <div class="m-2 p-2 rounded-md bg-error-100">bg-error-100</div>
+        <div class="m-2 p-2 rounded-md bg-error-200">bg-error-200</div>
+        <div class="m-2 p-2 rounded-md bg-error-300">bg-error-300</div>
+        <div class="m-2 p-2 rounded-md bg-error-400">bg-error-400</div>
+        <div class="m-2 p-2 rounded-md bg-error-500">bg-error-500</div>
+        <div class="m-2 p-2 rounded-md bg-error-600">bg-error-600</div>
+        <div class="m-2 p-2 rounded-md bg-error-700">bg-error-700</div>
+        <div class="m-2 p-2 rounded-md bg-error-800">bg-error-800</div>
+        <div class="m-2 p-2 rounded-md bg-error-900">bg-error-900</div>
+      </div>
+    </div>
+
+    <h1 class="text-4xl underline font-medium my-4">Typography</h1>
+
+    <h1 class="my-2 text-6xl">Title H1</h1>
+    <h2 class="my-2 text-5xl">Title H2</h2>
+    <h3 class="my-2 text-4xl">Title H3</h3>
+    <h4 class="my-2 text-3xl">Title H4</h4>
+    <h5 class="my-2 text-2xl">Title H5</h5>
+    <h6 class="my-2 text-xl">Title H6</h6>
+  </div>
+</template>